values went from double to float

+++ ApiClient/Models/PlayerState.cs
@@ -6,14 +6,14 @@
         public string Id { get; set; }
         public int MutationPoints { get; set; }
         public bool Human { get; set; }
-        public double TopLeftGrowthChance { get; set; }
-        public double TopGrowthChance { get; set; }
-        public double TopRightGrowthChance { get; set; }
-        public double RightGrowthChance { get; set; }
-        public double BottomRightGrowthChance { get; set; }
-        public double BottomGrowthChance { get; set; }
-        public double BottomLeftGrowthChance { get; set; }
-        public double LeftGrowthChance { get; set; }
+        public float TopLeftGrowthChance { get; set; }
+        public float TopGrowthChance { get; set; }
+        public float TopRightGrowthChance { get; set; }
+        public float RightGrowthChance { get; set; }
+        public float BottomRightGrowthChance { get; set; }
+        public float BottomGrowthChance { get; set; }
+        public float BottomLeftGrowthChance { get; set; }
+        public float LeftGrowthChance { get; set; }
         public int DeadCells { get; set; }
         public int LiveCells { get; set; }
         public int RegeneratedCells { get; set; }
@@ -22,11 +22,11 @@
         public int RegenerationSkillLevel { get; set; }
         public int BuddingSkillLevel { get; set; }
         public int MycotoxinsSkillLevel { get; set; }
-        public double ApoptosisChance { get; set; }
-        public double StarvedCellDeathChance { get; set; }
-        public double MutationChance { get; set; }
-        public double RegenerationChance { get; set; }
-        public double MycotoxinFungicideChance { get; set; }
+        public float ApoptosisChance { get; set; }
+        public float StarvedCellDeathChance { get; set; }
+        public float MutationChance { get; set; }
+        public float RegenerationChance { get; set; }
+        public float MycotoxinFungicideChance { get; set; }
         public string Status { get; set; }
     }
 }
